# docs/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "BlueMoon Admin API",
      "description": "API Documentation for BlueMoon Apartment Management System",
      "version": "1.2.1",
      "contact": {
        "name": "BlueMoon Admin Team",
        "email": "admin@bluemoon.com"
      },
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "servers": [
      {
        "url": "http://localhost:5000",
        "description": "Local Development Server"
      }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
...
        "Fee": {
          "type": "object",
          "properties": {
            "_id": { "type": "string" },
            "title": { "type": "string" },
            "description": { "type": "string" },
            "type": {
              "type": "string",
              "enum": ["Service", "Contribution", "Utility"]
            },
            "amount": { "type": "number" },
            "unit": {
              "type": "string",
              "enum": ["hộ", "khẩu", "m2", "kWh", "m3"]
            },
            "isActive": { "type": "boolean" }
          }
        }
      }
    },
    "security": [{ "bearerAuth": [] }],
    "paths": {
...
      "/api/fees": {
        "get": {
          "summary": "Get Fees List",
          "parameters": [
            { "in": "query", "name": "type", "schema": { "type": "string" } },
            { "in": "query", "name": "limit", "schema": { "type": "integer" } }
          ],
          "operationId": "getFees",
          "responses": {
            "200": { "description": "List of fees" },
            "400": { "description": "Bad request" }
          }
        },
        "post": {
          "summary": "Create Fee",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Fee" }
              }
            }
          },
          "operationId": "createFee",
          "responses": {
            "201": { "description": "Fee created" },
            "400": { "description": "Validation error" }
          }
        }
      },
...
    }
  }